Move isSerialized to class text
